♻ Refactor `logic` module into package
