+++ src/logic/__init__.py
@@ -1,0 +1,3 @@
+from .base import process_event
+
+__all__ = ('process_event',)
